requirements.txt
# Apache Airflow Providers
apache-airflow-providers-postgres==5.6.1
apache-airflow-providers-apache-spark==4.1.3

# PySpark and Data Processing
pyspark==3.4.1
pyspark[sql]==3.4.1

# Database Connectivity
psycopg2-binary==2.9.7
SQLAlchemy==1.4.49

# Data Analysis and Visualization
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2

# File Processing
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development and Testing
pytest==7.4.0
black==23.7.0
flake8==6.0.0 